Extended weapon support

+++ src/Components/AnimationFrame.h
@@ -3,5 +3,5 @@
 struct AnimationFrame
 {
     long tileID{};
-    long duration{};
+    float duration{};
 };
